Reshuffle api.py into module
