+++ bin/debug-api
@@ -1,0 +1,8 @@
+#!/usr/bin/env python
+
+import sys
+from os.path import abspath, split, join
+sys.path.append(join(split(abspath(__file__))[0], '..'))
+
+from elevation import api
+api.app.run(debug=True)
